Add our improved check_xpect script
The original was only capable of checking that all expectations defined
in an account were in a successful state. Now you can add an optional
parameter specifying the ID of a specific expectation. This can be very
useful for setting up more granular Nagios checks.

Additionally, some minor improvements to the source code formatting have
been made as well as some durability improvements.